fix: refactor printFileAclForUser sample into its own file
